added basic lighting, redid terrain model generation to generate actual normals for vertices, added a grass texture

+++ src/game.h
@@ -35,7 +35,7 @@ enum TileResource {
 };
 
 struct Tile {
-	i32 elevation;
+	f32 elevation;
 	TileResource resources[3];
 };
